+++ hw-ledger/bundle.ts
@@ -1,11 +1,11 @@
 // Copyright 2017-2022 @polkadot/hw-ledger authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import type { ResponseBase, SubstrateApp } from 'https://esm.sh/@zondax/ledger-substrate@0.32.0';
+import type { ResponseBase, SubstrateApp } from 'https://esm.sh/@zondax/ledger-substrate@0.33.0';
 import type { AccountOptions, LedgerAddress, LedgerSignature, LedgerTypes, LedgerVersion } from './types.ts';
 
-import { transports } from 'https://deno.land/x/polkadot@0.0.6/hw-ledger-transports/mod.ts';
-import { u8aToBuffer } from 'https://deno.land/x/polkadot@0.0.6/util/mod.ts';
+import { transports } from 'https://deno.land/x/polkadot/hw-ledger-transports/mod.ts';
+import { u8aToBuffer } from 'https://deno.land/x/polkadot/util/mod.ts';
 
 import { LEDGER_DEFAULT_ACCOUNT, LEDGER_DEFAULT_CHANGE, LEDGER_DEFAULT_INDEX, LEDGER_SUCCESS_CODE } from './constants.ts';
 import { ledgerApps } from './defaults.ts';