+++ util/array/index.ts
@@ -10,4 +10,5 @@ export { arrayFilter } from './filter.ts';
 export { arrayFlatten } from './flatten.ts';
 export { arrayRange } from './range.ts';
 export { arrayShuffle } from './shuffle.ts';
+export { arrayUnzip } from './unzip.ts';
 export { arrayZip } from './zip.ts';