+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master
 
+- hw-{ledger, ledger-transports}, keyring, networks, util, util-crypto, x-{bigint, fetch, global, randomvalues, textdecoder, textencoder, ws} 10.1.1
 - wasm-{bridge, crypto, crypto-{asmjs, init, wasm}, util} 6.3.1
 
 ## 0.0.6 Jul 16, 2022